Modify chains security context config
The goal is to avoid needing to run the `oc adm policy
add-scc-to-user ...` command to get chains working in openshift.

IIUC that command is a workaround for some some security
configuration that doesn't work well with openshift. With this
change we remove the configuration so the workaround is not
required.

Also, since we're no longer setting the project in
bootstrap-cluster.sh, let's remove the oc project default.

This change was suggested by @Michkov. I'm not sure if there will be
any other impacts or security concerns, but so far it seems to work
okay.

See also:
* https://github.com/tektoncd/chains/commit/727ff49d6dca9d87d5eb5b86710cdc182acb60b8